requirements.txt
spotipy
python-dotenv
google-generativeai
Flask
